refactor(react-gui): split the drag field into its three gestures
DragNumericField was 851 lines holding three interaction models at once:
a pointer-lock body drag, an auto-repeating arrow press, and a text
editor with its own keyboard stepping. Each has invariants the others do
not -- the drag's listeners must stay reference-stable or teardown leaks
one; the press must accumulate its own total because the parent may not
have echoed the value yet; the editor must not commit on the blur that an
arrow press causes -- and reading any of them meant scrolling past the
other two.

They are now three hooks over a shared core. The core is small and named:
the field has one mode and one draft between the gestures, because a
press that never moves becomes an edit, an arrow press leaves edit mode
first, and a blur has to know a press is running. Those hand-offs are
the coupling; everything else -- drag bookkeeping, press timers, the
key-step accumulator and caret restore -- is private to its hook.

The interaction logic moved verbatim; the only edits the extraction
forced are exported declarations, parameters where there were closure
captures, `isPressing()` where the editor read `pressRef`, and one
unmount cleanup becoming two (a press and a body drag cannot overlap, so
at most one still cancels).

Total lines go up, not down -- 851 to 1,127 -- and that is the trade: the
shared state now has to be declared and passed rather than being in
scope. What it buys is that no file is over 350 lines, each states what
it may not break, and the 38 existing tests pass unchanged, which is what
makes the move checkable at all.
